Fix visualization axis helpers
