+++ MANIFEST.in
@@ -1,3 +1,4 @@
 include README.md
 include LICENSE
 recursive-include bids_manager/miscellaneous/images *
+recursive-include bids_manager/visualization *.py
