+++ pyproject.toml
@@ -51,7 +51,7 @@ post-conv-renamer = "bids_manager.post_conv_renamer:main"
 fill-bids-ignore = "bids_manager.fill_bids_ignore:main"
 
 [tool.setuptools]
-packages = ["bids_manager"]
+packages = ["bids_manager", "bids_manager.visualization"]
 include-package-data = true
 
 [tool.setuptools.package-data]
